+++ .gitignore
@@ -4,3 +4,4 @@
 !/lib/.gitkeep
 .vscode/*
 */.vscode/*
+.editorconfig
