+++ CMakeLists.txt
@@ -5,16 +5,15 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_CXX_EXTENSIONS OFF)
 
-add_subdirectory(lib/Server)
-# add_subdirectory(lib/Handlers)
-add_subdirectory(lib/ML)
+add_subdirectory(lib/server)
+add_subdirectory(lib/computer_vision)
+add_subdirectory(lib/database_handler)
+add_subdirectory(lib/tg_bot)
 
 add_executable(${PROJECT_NAME}  src/main.cpp)
 
 
-target_link_libraries(${PROJECT_NAME} Server ML)
-# target_link_libraries(${PROJECT_NAME} Server Handlers ML)
-
+target_link_libraries(${PROJECT_NAME} server computer_vision database_handler tg_bot)
 
 # target_compile_options(${PROJECT_NAME} PRIVATE -Wall -Wextra -pedantic)
