--- lib/Handlers/src/GetQueueCountHandler.cpp
@@ -1,4 +1,0 @@
-#include "GetQueueCountHandler.hpp"
-#include <random>
-
-GetQueueCountHandler::handle_request
