+++ lib/computer_vision/include/model.hpp
