+++ lib/computer_vision/src/model.cpp
