+++ lib/database_handler/include/api.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <iostream>
+
+void database_handler();
