+++ lib/database_handler/src/api.cpp
@@ -1,0 +1,5 @@
+#include "api.hpp"
+
+void database_handler(){
+    std::cout << "database_handler printing" << std::endl;
+}
