+++ lib/server/CMakeLists.txt
@@ -3,7 +3,7 @@ aux_source_directory(src SERVER_SOURCES)
 
 # Задание цели
 # add_library(Server ${SERVER_SOURCES})
-add_library(Server  
+add_library(server  
                     src/error.cpp                    
                     src/session.cpp                    
                     src/listener.cpp                    
@@ -15,4 +15,4 @@ add_library(Server
 )
 
 # Покдлючение всех hpp файлов в директории include/
-target_include_directories(Server PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
+target_include_directories(server PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include)
