+++ lib/tg_bot/include/bot.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <iostream>
+
+void bot();
