+++ lib/tg_bot/src/bot.cpp
@@ -1,0 +1,5 @@
+#include "bot.hpp"
+
+void bot(){
+    std::cout << "Tg bot printing" << std::endl;
+}
